Add more users and info.txt

+++ Rtippo6Sivkov/Controllers/HomeController.cs
@@ -49,24 +49,73 @@ public class HomeController : Controller
         return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
     }
 
-    public async Task<IActionResult> CreateAdminUser(
+    public async Task<IActionResult> CreateUsers(
         [FromServices] UserManager<AppUser> userManager,
         [FromServices] ApplicationDbContext dbContext)
     {
-        _logger.LogInformation("Creating Admin user with password '__@Dm1n__'");
-        var user = new AppUser("Admin")
+        var guestCreationResult = await CreateUser("Guest", "__Gu3sT__", "Москва", "Гость");
+        if (!guestCreationResult.Succeeded)
         {
-            Email = "admin@localhost.com"
-        };
-        var userToDelete = dbContext.Users.Single(x => x.UserName == "Admin");
-        await userManager.DeleteAsync(userToDelete);
-        var result = await userManager.CreateAsync(user, "__@Dm1n__");
-        
-        if (result.Succeeded)
-        {
-            return Ok(user);
+            return Conflict(new { guestCreationResult.Errors });
         }
 
-        return Conflict(new { result.Errors });
+        var adminCreationResult = await CreateUser("Admin", "__@Dm1n__", "Москва", "Администратор");
+        if (!adminCreationResult.Succeeded)
+        {
+            return Conflict(new { adminCreationResult.Errors });
+        }
+
+        var vetCuratorCreationResult = await CreateUser("VetCurator", "__V3tCur@t0r__", "Тюмень", "Куратор Ветслужбы");
+        if (!vetCuratorCreationResult.Succeeded)
+        {
+            return Conflict(new { vetCuratorCreationResult.Errors });
+        }
+
+        var omsuCuratorCreationResult = await CreateUser("OmsuCurator", "__0msUCur@t0r__", "Тюмень", "Куратор ОМСУ");
+        if (!omsuCuratorCreationResult.Succeeded)
+        {
+            return Conflict(new { omsuCuratorCreationResult.Errors });
+        }
+
+        var vetOperatorCreationResult =
+            await CreateUser("VetOperator", "__V3t0p3r@T0r__", "Тюмень", "Оператор Ветслужбы");
+        if (!vetOperatorCreationResult.Succeeded)
+        {
+            return Conflict(new { vetOperatorCreationResult.Errors });
+        }
+
+        var omsuOperatorCreationResult = await CreateUser("OmsuOperator", "__0msU0p3r@T0r__", "Тюмень", "Оператор ОМСУ");
+        if (!omsuOperatorCreationResult.Succeeded)
+        {
+            return Conflict(new { omsuOperatorCreationResult.Errors });
+        }
+
+        return Ok();
+
+        async Task<IdentityResult> CreateUser(string username, string password, string localityName, string roleName)
+        {
+            _logger.LogInformation(
+                "@Creating '{Username}' user with password '{Password}', locality {Locality} and role {RoleName}",
+                username, password, localityName, roleName);
+
+            var userToDelete = dbContext.Users.SingleOrDefault(x => x.UserName == username);
+            if (userToDelete != null)
+            {
+                await userManager.DeleteAsync(userToDelete);
+            }
+
+            var localityEntity = _dbContext.Localities.SingleOrDefault(x => x.Name.ToLower() == localityName);
+            var roleEntity = _dbContext.AppUserRoles.SingleOrDefault(x => x.Name.ToLower() == roleName);
+
+            var userToCreate = new AppUser(username)
+            {
+                Email = $"{username}@localhost.com",
+                Locality = localityEntity ?? throw new Exception($"Locality with name {localityName} not found"),
+                Role = roleEntity ?? throw new Exception($"Role with name {roleName} not found")
+            };
+            var result = await userManager.CreateAsync(userToCreate, password);
+
+            return result;
+        }
     }
 }
